+++ src/app/actions.rs
@@ -2968,6 +2968,7 @@ impl AppState {
                 Vec::new()
             }
             AppEvent::ProjectStatusRefreshed { results: _ } => Vec::new(),
+            AppEvent::SessionStatusRefreshed { results: _ } => Vec::new(),
             AppEvent::WorktreeAddFinished(_) => Vec::new(),
             AppEvent::WorktreeRemoveFinished(_) => Vec::new(),
             AppEvent::PluginCommandFinished { .. } => Vec::new(),